+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 #### 0.0.1.alpha.1
+
+* add metrics-golang-stats-api